pyproject.toml
[project]
name = "opamp-server-py"
version = "0.1.0"
description = ""
authors = [
    {name = "qdeli"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "websocket (>=0.2.1,<0.3.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "protobuf-to-pydantic[all] (>=0.3.1.1,<0.4.0.0)",
    "uuid7 (>=0.1.0,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

